pyproject.toml
[tool.poetry]
name = "llamaindex-retrieval-plugin"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "server"}]

[tool.poetry.dependencies]
python = "^3.9"
aiohttp = "3.9.1"
aiosignal = "1.3.1"
annotated-types = "0.6.0"
anyio = "4.2.0"
arrow = "1.3.0"
arxiv = "2.1.0"
async-timeout = "4.0.3"
attrs = "23.2.0"
authlib = "1.3.0"
beautifulsoup4 = "4.12.2"
certifi = "2023.11.17"
cffi = "1.16.0"
charset-normalizer = "3.3.2"
click = "8.1.7"
cryptography = "41.0.7"
dataclasses-json = "0.6.3"
deprecated = "1.2.14"
distro = "1.9.0"
docx2txt = "0.8"
exceptiongroup = "1.2.0"
fastapi = "0.109.0"
feedparser = "6.0.10"
frozenlist = "1.4.1"
fsspec = "2023.12.2"
greenlet = "3.0.3"
h11 = "0.14.0"
httpcore = "1.0.2"
httpx = "0.26.0"
idna = "3.6"
iniconfig = "2.0.0"
joblib = "1.3.2"
llama-index = "0.9.30"
loguru = "0.7.2"
lxml = "5.1.0"
marshmallow = "3.20.2"
multidict = "6.0.4"
mypy-extensions = "1.0.0"
nest-asyncio = "1.5.8"
networkx = "3.2.1"
nltk = "3.8.1"
numpy = "1.26.3"
openai = "1.7.2"
packaging = "23.2"
pandas = "2.1.4"
pillow = "10.2.0"
pluggy = "1.3.0"
pycparser = "2.21"
pydantic = "2.5.3"
pydantic-core = "2.14.6"
pypdf2 = "3.0.1"
pytest = "7.4.4"
python-dateutil = "2.8.2"
python-multipart = "0.0.6"
python-pptx = "0.6.23"
pytz = "2023.3.post1"
regex = "2023.12.25"
requests = "2.31.0"
sgmllib3k = "1.0.0"
six = "1.16.0"
sniffio = "1.3.0"
soupsieve = "2.5"
sqlalchemy = "2.0.25"
starlette = "0.35.1"
tenacity = "8.2.3"
tiktoken = "0.5.2"
tomli = "2.0.1"
tqdm = "4.66.1"
types-python-dateutil = "2.8.19.20240106"
typing-inspect = "0.9.0"
typing-extensions = "4.9.0"
tzdata = "2023.4"
urllib3 = "2.1.0"
uvicorn = "0.25.0"
validators = "0.22.0"
weaviate-client = "3.26.0"
wrapt = "1.16.0"
xlsxwriter = "3.1.9"
yarl = "1.9.4"

[tool.poetry.scripts]
start = "server.main:start"
dev = "local_server.main:start"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
